add payment module
