+++ crates/fiber-lib/src/fiber/tests/mpp.rs
@@ -10,9 +10,9 @@ use crate::{
         },
         config::{CKB_SHANNONS, DEFAULT_TLC_EXPIRY_DELTA, PAYMENT_MAX_PARTS_LIMIT},
         features::FeatureVector,
-        graph::AttemptStatus,
         hash_algorithm::HashAlgorithm,
         network::{DebugEvent, SendPaymentCommand},
+        payment::AttemptStatus,
         types::{Hash256, PaymentDataRecord, PaymentHopData, PeeledOnionPacket, RemoveTlcReason},
         NetworkActorCommand, NetworkActorMessage, PaymentCustomRecords,
     },
