+++ crates/fiber-lib/src/tests/test_utils.rs
@@ -8,10 +8,10 @@ use crate::fiber::features::FeatureVector;
 use crate::fiber::gossip::get_gossip_actor_name;
 use crate::fiber::gossip::GossipActorMessage;
 use crate::fiber::graph::NetworkGraphStateStore;
-use crate::fiber::graph::PaymentSession;
-use crate::fiber::graph::PaymentStatus;
-use crate::fiber::graph::SessionRoute;
 use crate::fiber::network::*;
+use crate::fiber::payment::PaymentSession;
+use crate::fiber::payment::PaymentStatus;
+use crate::fiber::payment::SessionRoute;
 use crate::fiber::types::EcdsaSignature;
 use crate::fiber::types::FiberMessage;
 use crate::fiber::types::GossipMessage;
